requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
DAWG-Python==0.7.2
filelock==3.14.0
fsspec==2024.5.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.23.0
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pika==1.3.2
preshed==3.0.9
pydantic==2.7.1
pydantic_core==2.18.2
pymorphy3==2.0.1
pymorphy3-dicts-ru==2.4.417150.4580142
pymorphy3-dicts-uk==2.4.1.1.1663094765
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.3
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.3
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
typer==0.9.4
typing_extensions==4.11.0
uk-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/uk_core_news_sm-3.7.0/uk_core_news_sm-3.7.0-py3-none-any.whl
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
